+++ tests/unit/compile/cases/tags/instruction/block/01.basic.t
@@ -1,3 +1,3 @@
 Lorem {{ %block foo }} ipsum
 ---- OK ----
-Lorem <?php echo $runtime->block('foo') ?> ipsum
+Lorem <?php $runtime->renderBlock('foo') ?> ipsum